Fix worst strategy: Its winrate is 53.3%, not 40%
The bug in the old worst strategy was:
Assume there is only 1 fruit on 1 tree, and 0 on the 3 others.
You roll a basket and invoke the worst strategy.
First, the lone fruit is removed.
Second, you hang 10 new fruit on the first tree.

This makes the round nearly unwinnable, and was the reason for a
dent in the win rate (to 40%) instead of 53.3%.

Now, the win rates of all 3 strategies match what I determined
with my completely unrelated implementation before I found yours. :-)